feat: replace custom tracing with tokio's tracing integration
After some experiments using our custom tracing implementation, I
reached the conclusion that it isn't worth the extra complexity. Most
of the ecossystem is using tokio's tracing implementation, and the Rust
OpenTelemetry WG are evaluating replacing their implementation with only
the tracing layer[1].

Thus, I decided to replace it with a tracing integration setup. The
setup is pretty standard, but the implementation uses a custom Layer
to pass data from tracing to OpenTelemetry, continuing to use the
background worker to do most of the heavy lifting. This makes the
hot path that runs in the application loop to be more efficient.

The implementation also uses a custom context to allow for faster
retrieval of tracing information for propagation.

The result is that kiso's users don't have to worry about OpenTelemetry
crates unless they need dynamic attributes or links. Everything else is
handled by the tracing crate.

This commit contains only the necessary code for the migration. There
are still some things to sort out, and primarily performance
improvement to implement. These will be done in other patches as this
one is already too big to properly review.

[1]: https://github.com/open-telemetry/opentelemetry-rust/issues/1571